Update Firebase SDK to 8.9.0
