--- ios/platform/FirebasePerformance.framework/Headers/FirebasePerformance.h
@@ -1,4 +1,0 @@
-#import "FIRHTTPMetric.h"
-#import "FIRPerformance.h"
-#import "FIRPerformanceAttributable.h"
-#import "FIRTrace.h"
