--- ios/platform/FirebasePerformance.framework/Modules/module.modulemap
@@ -1,13 +1,0 @@
-framework module FirebasePerformance {
-  umbrella header "FirebasePerformance.h"
-  export *
-  module * { export *}
-  link "c++"
-  link "sqlite3"
-  link "z"
-  link framework "CoreTelephony"
-  link framework "QuartzCore"
-  link framework "Security"
-  link framework "StoreKit"
-  link framework "SystemConfiguration"
-  link framework "UIKit"}
