+++ ios/platform/FirebasePerformance.xcframework/ios-arm64_x86_64-maccatalyst/FirebasePerformance.framework/Modules/module.modulemap
@@ -1,0 +1,11 @@
+framework module FirebasePerformance {
+umbrella header "FirebasePerformance.h"
+export *
+module * { export * }
+  link framework "CoreTelephony"
+  link framework "Foundation"
+  link framework "Security"
+  link framework "SystemConfiguration"
+  link framework "UIKit"
+  link "z"
+}
