+++ ios/platform/FirebasePerformance.xcframework/tvos-arm64/FirebasePerformance.framework/Modules/module.modulemap
@@ -1,0 +1,10 @@
+framework module FirebasePerformance {
+umbrella header "FirebasePerformance.h"
+export *
+module * { export * }
+  link framework "Foundation"
+  link framework "Security"
+  link framework "SystemConfiguration"
+  link framework "UIKit"
+  link "z"
+}
