--- ios/platform/FirebaseSwizzlingUtilities.framework/Modules/module.modulemap
@@ -1,4 +1,0 @@
-framework module FirebaseSwizzlingUtilities {
-  export *
-  module * { export *}
-  link "c++"}
